+++ sdk/app_cpu1/common/sys/scheduler.h
@@ -44,7 +44,7 @@ void scheduler_tick(void);
 void scheduler_run(void);
 
 void scheduler_tcb_init(
-    task_control_block_t *tcb, task_callback_t callback, void *callback_arg, const char *name, uint32_t interval_usec);
+    task_control_block_t *tcb, task_callback_t callback, void *callback_arg, const char *name, double interval_usec);
 int scheduler_tcb_register(task_control_block_t *tcb);
 int scheduler_tcb_register_high_priority(task_control_block_t *tcb);
 int scheduler_tcb_unregister(task_control_block_t *tcb);
